show all indicator categories on the main menu
- minor refactoring of stuff, putting all indicator mixins in their own
  module
